Improved Confirm Buttons in Power Skin
I wanted these buttons to have a very minimal "confirm" feature, so that they couldn't be triggered by an accidental single click. But having the original button hide and showing a confirm button in it's place meant that, if the user changed their mind or if this was indeed an accidental click, the confirm button just stays there indefinitely, and is now itself prone to the original accidental click problem. I fixed this by having the button reset to the original state when the user's mouse moves outside its bounds.