Add Bar Meter to Recycle Bin Skin
I wanted a bar here to make this Skin a more visually similar to the others. I wrote out the Meters and Calc necessary to make the bar show the percentage of used disk space comprised of the contents of the recycle bin, but even with my disks as full as they are, a gigabyte worth of files in the bin barely shows anything on the bar. So I'm just going to run with a binary bar instead -- anything in the bin, and the bar is full, otherwise the bar is empty.