Remove Decimal from All Temperature Meters
In an effort to keep these temperature meters relatively consistent in width, I implemented an IfCondition to remove the decimal if the temperature was greater than or equal to 100... But really, this degree of specificity is overkill -- Knowing the whole number is good enough. So, I removed all the decimal declarations instead, and the result is a lot more minimalist.